+++ core/api_container/server/bulk_command_execution_engine/v0_bulk_command_execution/v0_bulk_commands_processor.go
@@ -1,4 +1,4 @@
-package v0_bulk_command_api
+package v0_bulk_command_execution
 
 import (
 	"context"